Its funny, didn't get "parole", and don't care. I spent my youth unearthing our "non functioning society", and my prime figuring out how to ease the
transition to functioning one with rates down implementing Nature's currency. Now people are (finally) talking about inflation and rising cost of living all over the world, which could have

propelled rd to be the beacon of frontline innovation needed to transition to a functioning society that I wanted it to be. Instead I met soul

crushing antagonism, both active and passive. I dreamt up techvolution to be the narrative that explains all this but can't work up the motivation because the path to actualization is yet another

 face saying innovation is great "except in my backyard, oh btw Biden/Trump are evil". All the while, technologists

 aren't ready to protect innovation inside a cohesive narrative because they aren't ready to climb the ladder

away from consumerism proactively before IA philosophies disintegrate.
 Ashses to ashses, dust to dust I guess.